Extract mapStyle
