+++ app/components/cal/map.vue
@@ -72,8 +72,8 @@ import { useToggle } from '@vueuse/core'
 import { type CensusGeography, type Stop, stopToStopCsv, type Route, routeToRouteCsv } from '~~/src/tl'
 import type { Marker } from 'maplibre-gl'
 import type { Bbox, Feature, Point, PopupFeature, MarkerFeature, MarkerDragEvent, ChoroplethClassification, ClusterMemberInfo } from '~~/src/core'
-import { colors, categoricalColors, routeTypeNames, flexColors, createCategoryColorScale } from '~~/src/core'
-import type { ScenarioFilterResult, StopCluster } from '~~/src/scenario'
+import { categoricalColors, routeTypeNames, flexColors, createCategoryColorScale } from '~~/src/core'
+import { buildStyleData, type Matcher, type ScenarioFilterResult, type StopCluster } from '~~/src/scenario'
 
 const emit = defineEmits<{
   setMapExtent: [value: Bbox]
@@ -590,184 +590,14 @@ const agencyData = computed((): AgencyData[] => {
 const agencyColorScale = computed(() =>
   createCategoryColorScale(agencyData.value.map(a => String(a.numericId)), categoricalColors))
 
-// Depending on the data display, set up matcher rules to choose a styling.
-// Matchers should run in the order that they are added to the rules array.
-type MatchFunction = (x: Stop | Route) => boolean
-
-interface Matcher {
-  label: string
-  color: string
-  match: MatchFunction
-}
-const styleData = computed((): Matcher[] => {
-  const routeLookup = new Map<number, Route>()
-  for (const route of props.scenarioFilterResult?.routes || []) {
-    routeLookup.set(route.id, route)
-  }
-
-  const stopLookup = new Map<number, Stop>()
-  for (const stop of props.scenarioFilterResult?.stops || []) {
-    stopLookup.set(stop.id, stop)
-  }
-
-  const routeStopLookup = new Map<number, number[]>()
-  for (const stop of props.scenarioFilterResult?.stops || []) {
-    for (const rs of stop.route_stops) {
-      const rid = rs.route.id
-      const stops = routeStopLookup.get(rid) || []
-      stops.push(stop.id)
-      routeStopLookup.set(rid, stops)
-    }
-  }
-
-  // Style based on AGENCY
-  function getAgencyMatcher (val: string): MatchFunction {
-    return (v: any) => {
-      if (v.__typename === 'Stop') {
-        return (v as Stop).route_stops.some((rs: any) => rs.route.agency?.agency_id === val)
-      } else if (v.__typename === 'Route') {
-        return (v as Route).agency?.agency_id === val
-      }
-      return false
-    }
-  }
-
-  // Style based on ROUTE MODE
-  function getModeMatcher (val: number): MatchFunction {
-    return (v: any) => {
-      if (v.__typename === 'Stop') {
-        // Filter out routes with null/undefined route_type to avoid false matches
-        // Also check that route data exists (may still be loading)
-        const validRoutes = (v as Stop).route_stops.filter((rs: any) => rs.route && rs.route.route_type != null)
-        // If no valid routes, don't match any mode (routes may still be loading)
-        if (validRoutes.length === 0) {
-          return false
-        }
-        // Match if ANY route at this stop has this mode (not every)
-        // This allows multi-modal stops to match their highest-priority mode
-        return validRoutes.some((rs: any) => rs.route.route_type === val)
-      } else if (v.__typename === 'Route') {
-        // For routes, also check for null/undefined
-        if ((v as Route).route_type == null || (v as Route).route_type == undefined) {
-          return false
-        }
-        return (v as Route).route_type === val
-      }
-      return false
-    }
-  }
-
-  // Style based on ROUTE FREQUENCY
-  function getRouteFrequencyMatcher (val: number): MatchFunction {
-    return (v: any) => {
-      if (v.__typename === 'Stop') {
-        return (v as Stop).route_stops.some((rs: any) => {
-          const route = routeLookup.get(rs.route.id)
-          const headway = route?.average_frequency || -1
-          return headway >= val * 60
-        })
-      } else if (v.__typename === 'Route') {
-        const headway = (v as Route).average_frequency || -1
-        return headway >= val * 60
-      }
-      return false
-    }
-  }
-
-  // Style based on STOP VISIT COUNT
-  function getStopVisitMatcher (val: number): MatchFunction {
-    return (v: any) => {
-      if (v.__typename === 'Stop') {
-        const count = (v as Stop).visits?.total.visit_average || -1
-        return count >= val
-      } else if (v.__typename === 'Route') {
-        const stopIds = routeStopLookup.get((v as Route).id) || []
-        return stopIds.some((sid) => {
-          const stop = stopLookup.get(sid)
-          return (stop?.visits?.total.visit_average || -1) >= val
-        })
-      }
-      return false
-    }
-  }
-
-  // Agencies use the wider 10-color categorical palette (agencyColorScale),
-  // not the 6-color route palette.
-  function getAgencyMatchers (): Matcher[] {
-    const rules: Matcher[] = []
-    const agencies = agencyData.value || []
-    const scale = agencyColorScale.value
-    for (let i = 0; i < Math.min(agencies.length, categoricalColors.length); i++) {
-      const agency = agencies[i]
-      if (agency) {
-        rules.push({ label: agency.name ?? '', color: scale(String(agency.numericId)), match: getAgencyMatcher(agency.id ?? '') })
-      }
-    }
-    return rules
-  }
-
-  // Generate a set of MODE MATCHERS (static)
-  function getModeMatchers (): Matcher[] {
-    const rules: Matcher[] = []
-    const modes = [...routeTypeNames.keys()]
-    for (let i = 0; i < Math.min(modes.length, maxColor); i++) {
-      const mode = modes[i]
-      if (mode !== undefined) {
-        const label = routeTypeNames.get(mode) || 'Unknown'
-        rules.push({ label: label, color: colors[i]!, match: getModeMatcher(mode) })
-      }
-    }
-    return rules
-  }
-
-  // Generate a set of ROUTE FREQUENCY MATCHERS (static)
-  function getRouteFrequencyMatchers (): Matcher[] {
-    const rules: Matcher[] = []
-    rules.push({ label: '40+ mins', color: colors[0], match: getRouteFrequencyMatcher(40) })
-    rules.push({ label: '30-39 mins', color: colors[1], match: getRouteFrequencyMatcher(30) })
-    rules.push({ label: '20-29 mins', color: colors[2], match: getRouteFrequencyMatcher(20) })
-    rules.push({ label: '10-19 mins', color: colors[3], match: getRouteFrequencyMatcher(10) })
-    rules.push({ label: '0-9 mins', color: colors[4], match: getRouteFrequencyMatcher(0) })
-    return rules
-  }
-
-  // Generate a set of STOP VISIT MATCHERS (static)
-  function getStopVisitMatchers (): Matcher[] {
-    const rules: Matcher[] = []
-    rules.push({ label: '100+ visits', color: colors[0], match: getStopVisitMatcher(100) })
-    rules.push({ label: '50-100 visits', color: colors[1], match: getStopVisitMatcher(50) })
-    rules.push({ label: '20-50 visits', color: colors[2], match: getStopVisitMatcher(20) })
-    rules.push({ label: '10-20 visits', color: colors[3], match: getStopVisitMatcher(10) })
-    rules.push({ label: '0-9 visits', color: colors[4], match: getStopVisitMatcher(0) })
-    return rules
-  }
-
-  // Reserve an extra color for "other", if needed. Agency mode draws from the
-  // wider categorical palette, so its "Other" bucket only kicks in past that.
-  const maxColor = colors.length - 1
-  const rules: Matcher[] = []
-
-  let otherThreshold = maxColor
-  if (dataDisplayMode.value === 'Agency') {
-    rules.push(...getAgencyMatchers())
-    otherThreshold = categoricalColors.length
-  } else if (dataDisplayMode.value === 'Transit mode') {
-    rules.push(...getModeMatchers())
-  } else if (dataDisplayMode.value === 'Route frequency') {
-    rules.push(...getRouteFrequencyMatchers())
-  } else if (dataDisplayMode.value === 'Stop visits') {
-    rules.push(...getStopVisitMatchers())
-  } else if (dataDisplayMode.value === 'Service area') {
-    // report-only mode; no map color rules
-  }
-
-  // If we used all colors (or no colors), add a catchall "other" rule
-  if (rules.length >= otherThreshold || rules.length === 0) {
-    rules.push({ label: 'Other', color: '#000', match: _ => true })
-  }
-
-  return rules
-})
+// Matcher rules color stops/routes by the active data-display mode. The pure
+// builder (and the Matcher type) live in src/scenario/map-style.ts.
+const styleData = computed((): Matcher[] => buildStyleData({
+  scenarioFilterResult: props.scenarioFilterResult,
+  dataDisplayMode: dataDisplayMode.value,
+  agencies: agencyData.value,
+  agencyColorScale: agencyColorScale.value,
+}))
 
 // Selectable geography features for click-to-select in adminBoundary mode.
 // Three visual states:
